feat: added seeding data for created user
