+++ .gitignore
@@ -6,4 +6,3 @@ node_modules
 .env
 /dev
 /server
-**/images
