+++ app/utils/ddb/seeding/seed.ts
@@ -1,0 +1,4 @@
+import { createTestUser } from "./data";
+import { seedData } from "./functions";
+
+seedData(createTestUser(), true);
