+++ package.json
@@ -13,6 +13,7 @@
     "destroy": "arc destroy --app remix-recipes --production --force"
   },
   "dependencies": {
+    "@aws-sdk/client-s3": "^3.583.0",
     "@getbrevo/brevo": "2.1.1",
     "@prisma/client": "5.13.0",
     "@remix-run/architect": "^2.9.2",
