+++ prefs.arc
@@ -1,2 +1,2 @@
 @sandbox-start
-npx tsx app/utils/ddb/seed.ts
+npx tsx app/utils/ddb/seeding/seed.ts
